Important fixes
