+++ src/components/dashboard/FaceVerification/index.js
@@ -1,92 +1,14 @@
 import React, { useCallback } from 'react'
-import { findKey, mapValues } from 'lodash'
 
 import SpinnerCheckMark from '../../common/animations/SpinnerCheckMark/SpinnerCheckMark'
 import Section from '../../common/layout/Section'
 
+import UserStorage from '../../../lib/gundb/UserStorage'
 import { useCurriedSetters } from '../../../lib/undux/GDStore'
 import goodWallet from '../../../lib/wallet/GoodWallet'
 import useZoomSDK from './hooks/useZoomSDK'
 import useZoomVerification from './hooks/useZoomVerification'
-import { ZoomSDKStatus, ZoomSessionStatus } from './sdk'
-
-const kindOfSessionIssuesMap = mapValues(
-  {
-    // All Zoom session result codes could be thrown if the
-    // camera isn't available or no camera access was confirmed
-    // in this case we're marking error as 'NotAllowedError'.
-    NotAllowedError: [
-      // camera permissions wasn't given
-      'UserCancelledWhenAttemptingToGetCameraPermissions',
-
-      // no camera available.
-      'CameraDoesNotExist',
-
-      // camera was not enabled.
-      'CameraNotEnabled',
-
-      // selected camera is not active
-      'CameraNotRunning',
-
-      // camera is busy because another ZoOm Session in progress.
-      'ZoomSessionInProgress',
-
-      // video initialization issues
-      'UnmanagedSessionVideoInitializationNotCompleted',
-      'VideoHeightOrWidthZeroOrUninitialized',
-      'VideoCaptureStreamNotActive',
-    ],
-
-    ForegoundLoosedError: [
-      // The ZoOm Session was cancelled due to the app being terminated, put to sleep, an OS notification,
-      // or the app was placed in the background (for web - tab was switched).
-      'ContextSwitch',
-    ],
-
-    // All Zoom sdk and session result codes could be thrown if device
-    // orientation isn't portrait or was changed during the session
-    // in this case we're marking error as 'DeviceOrientationError'.
-    DeviceOrientationError: [
-      // device orientation was changed during the ZoOm Session
-      'OrientationChangeDuringSession',
-
-      // device is in landscape mode
-      'LandscapeModeNotAllowed',
-    ],
-
-    // User has cancelled session by own decision
-    UserCancelled: [
-      // The user pressed the cancel button and did not complete the ZoOm Session.
-      'UserCancelled',
-
-      // The user pressed the cancel button during New User Guidance.
-      'UserCancelledFromNewUserGuidance',
-
-      // The user pressed the cancel button during Retry Guidance.
-      'UserCancelledFromRetryGuidance',
-
-      // The user cancelled out of the ZoOm experience while attempting to get camera permissions.
-      'UserCancelledWhenAttemptingToGetCameraPermissions',
-    ],
-  },
-  statusesKeys => statusesKeys.map(key => ZoomSessionStatus[key])
-)
-
-const kindOfSDKIssuesMap = mapValues(
-  {
-    // All Zoom sdk initialization result codes could be thrown
-    // if device orientation isn't portrait
-    // in this case we're marking error as 'DeviceOrientationError'.
-    DeviceOrientationError: [
-      // device is upside down
-      'DeviceInReversePortraitMode',
-
-      // device is in landscape mode
-      'DeviceInLandscapeMode',
-    ],
-  },
-  statusesKeys => statusesKeys.map(key => ZoomSDKStatus[key])
-)
+import { kindOfSDKIssue, kindOfSessionIssue } from './utils/kindOfTheIssue'
 
 const FaceVerification = ({ screenProps }) => {
   const [setIsCitizen] = useCurriedSetters(['isLoggedInCitizen'])
@@ -113,13 +35,10 @@ const FaceVerification = ({ screenProps }) => {
   // ZoomSDK session exception handler
   const exceptionHandler = useCallback(
     exception => {
-      // destructuring error object
-      const { message, code } = exception
-
       // the following code is needed for ErrorScreen component
       // could display specific error message corresponding to
       // the kind of issue (camera, orientation etc)
-      const kindOfTheIssue = findKey(kindOfSessionIssuesMap, statusCodes => statusCodes.includes(code))
+      const kindOfTheIssue = kindOfSessionIssue(exception)
 
       if ('UserCancelled' === kindOfTheIssue) {
         // If user has cancelled face verification by own
@@ -130,7 +49,7 @@ const FaceVerification = ({ screenProps }) => {
 
       if (kindOfTheIssue) {
         exception.name = kindOfTheIssue
-      } else if (message.startsWith('Duplicate')) {
+      } else if (exception.message.startsWith('Duplicate')) {
         exception.name = 'DuplicateFoundError'
       }
 
@@ -146,10 +65,10 @@ const FaceVerification = ({ screenProps }) => {
       // the following code is needed for ErrorScreen component
       // could display specific error message corresponding to
       // the kind of issue (camera, orientation etc)
-      const errorName = findKey(kindOfSDKIssuesMap, statusCodes => statusCodes.includes(exception.code))
+      const kindOfTheIssue = kindOfSDKIssue(exception)
 
-      if (errorName) {
-        exception.name = errorName
+      if (kindOfTheIssue) {
+        exception.name = kindOfTheIssue
       }
 
       // handling error
@@ -160,6 +79,7 @@ const FaceVerification = ({ screenProps }) => {
 
   // Using zoom verification hook, passing completion callback
   const startVerification = useZoomVerification({
+    enrollmentIdentifier: UserStorage.getFaceIdentifier(),
     onComplete: completionHandler,
     onError: exceptionHandler,
   })
