+++ src/components/dashboard/FaceVerification/sdk/index.native.js
@@ -34,12 +34,12 @@ export const ZoomSDK = new class {
     }
   }
 
-  async faceVerification() {
+  async faceVerification(enrollmentIdentifier) {
     try {
       // we're passing current JWT to the native code allowing it to call GoodServer for verification
       // unfortunately we couldn't pass callback which could return some data back to the native code
       // so it's only way to integrate Zoom on native - to reimplement all logic about calling server
-      const verificationStatus = await faceVerification(api.jwt)
+      const verificationStatus = await faceVerification(enrollmentIdentifier, api.jwt)
 
       return verificationStatus
     } catch (exception) {
